requirements.txt
fastapi
uvicorn
SQLAlchemy
pymongo
python-jose[cryptography]
passlib[bcrypt]
pydantic